+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 # [](https://github.com/Golang-Plus/math/compare/v1.2.0...v) (2017-04-21)
 
 
+### Bug Fixes
+
+* bugs in special cases ([3c4ca23](https://github.com/Golang-Plus/math/commit/3c4ca23))
+
+
+### Features
+
+* add more round functions ([907d20a](https://github.com/Golang-Plus/math/commit/907d20a))
+
+
+
+<a name=""></a>
+# [](https://github.com/Golang-Plus/math/compare/v1.2.0...v) (2017-04-21)
+
+
 ### Features
 
 * add more round functions ([907d20a](https://github.com/Golang-Plus/math/commit/907d20a))